% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{These packages are used in the .Rmd files so I am using
here to avoid warnings.}
\usage{
dummy()
}
\description{
These packages are used in the .Rmd files so I am using
here to avoid warnings.
}
