% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldilocks.R
\name{goldilocks}
\alias{goldilocks}
\title{This function takes the ui and server from \code{{goldilocks}} and runs it locally.}
\usage{
goldilocks()
}
\description{
This function takes the ui and server from \code{{goldilocks}} and runs it locally.
}
\examples{
 \dontrun{goldilocks()}
}
