% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldilocks.R
\name{goldilocksTwoArm}
\alias{goldilocksTwoArm}
\title{This function takes the ui and server from \code{{goldilocksTwoArm}} and runs it locally.}
\usage{
goldilocksTwoArm()
}
\description{
This function takes the ui and server from \code{{goldilocksTwoArm}} and runs it locally.
}
\examples{
 \dontrun{goldilocksTwoArm()}
}
