% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Ks.R
\name{make.Ks}
\alias{make.Ks}
\title{Title}
\usage{
make.Ks(M, vars, D, ToC, S)
}
\arguments{
\item{M}{a numeric vector of length one that is the number of times to practice randomizing, if this is too large it is hard to read the parallel coordinates plot. We recommend 300 which is the default.}

\item{vars}{a nested list that contains 5 things. In order: the name of the variable, the weight, the label for the plot, the min for the plot, and the max for the plot, all from input gui.}

\item{D}{a data frame input in the xlsx format from the gui}

\item{ToC}{If there are an odd number of units to randomize if it goes to treatment or control.}

\item{S}{}
}
\description{
Title
}
