% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_round.R
\name{make.order}
\alias{make.order}
\title{Interactive portion of the shiny app that
recolors the plot}
\usage{
make.order(column, data, M = 9)
}
\arguments{
\item{column}{a string that says which column should be used to color the parallel coordinates plot}

\item{data}{a dataframe made in make.Ks that need to be colored}

\item{M}{a number which is the Max number of colors allowed in YlGnBu palette from RColorBrewer is 9}
}
\description{
Interactive portion of the shiny app that
recolors the plot
}
