% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round1.R
\name{Round1}
\alias{Round1}
\title{Title}
\usage{
Round1(Max, Min, DC, K)
}
\arguments{
\item{Max}{numeric that is the maximum of the x value
from brushing the original. The limits of the original plot are
between 0 and 1 so Max is inside that interval and greater than Min.}

\item{DC}{numeric of where on the zoom plot the double click occurred}

\item{K}{numeric of the number of variables used for matching}
}
\value{
the variable that the plot should use to recolor the plot
}
\description{
@param Min numeric that is the minimum of the x value from brushing
the original plot. The limits of the original plot are between 0 and 1 so Min is
inside that interval
}
\examples{
Round1(Min = .2045, Max = .5234, DC = .43 , K = 20)


}
